backend/api/requirements.txt
# FastAPI Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.7

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Data
yfinance>=0.2.28
pandas>=2.1.0